+++ package.json
@@ -7,7 +7,7 @@
     "@emotion/styled": "11.3.0",
     "@material-ui/core": "^4.12.4",
     "@material-ui/icons": "4.11.2",
-    "@material-ui/lab": "^5.0.0-alpha.44",
+    "@material-ui/lab": "^4.0.0-alpha.60",
     "@mui/icons-material": "5.0.5",
     "@mui/material": "5.0.6",
     "@testing-library/jest-dom": "^5.16.5",
@@ -17,9 +17,11 @@
     "@types/node": "^16.18.13",
     "@types/react": "^17",
     "@types/react-dom": "^17",
+    "@types/redux": "3.6.0",
     "axios": "0.21.4",
     "react": "^17",
     "react-dom": "^17",
+    "react-redux": "7.2.5",
     "react-router-dom": "6",
     "react-scripts": "5.0.1",
     "react-use-localstorage": "3.5.3",
