--- src/paginas/services/Service.ts
@@ -1,15 +1,0 @@
-import axios from "axios";
-
-export const api = axios.create({
-    baseURL: 'https://vagalume.onrender.com'
-})
-
-
-export const cadastroUsuario = async(url: any,dados: any,setDado: any) => { 
-    const resposta = await api.post(url,dados)
-    setDado(resposta.data)
-}
-export const login = async(url: any,dados: any,setDado: any) => { 
-    const resposta = await api.post(url,dados)
-    setDado(resposta.data.token)
-  }
